+++ moonmcp/catalog.py
@@ -55,6 +55,7 @@ FAMILIES: OrderedDict[str, tuple[str, str, list[str]]] = OrderedDict([
             "cors_audit", "access_control_check", "authz_probe", "graphql_check",
             "ws_probe", "discover_parameters", "waf_detect", "takeover_check",
             "open_redirect", "trace_redirects", "crlf_probe", "vcs_exposure",
+            "git_forensics",
             "response_leak_probe", "reset_poison_probe", "path_bypass_probe",
             "debug_exposure",
             "screenshot", "browser_open", "browser_eval", "browser_interact", "cspp_probe",